variazione di tipo dell'array

+++ corso01/sketch35/sketch35.ino
@@ -16,7 +16,7 @@ int valore = 0;                   // valore restituito dall'analogRead()
 float temperatura = 0;            // variabile per conservare
                                   // la temperatura rilevata
 
-int misure[8];                    // array in cui memorizzare 8 valori
+float misure[8];                    // array in cui memorizzare 8 valori
                               
 void setup()
 {
